Fix: (farmer's side) real-time View All Bids; new bug
The buyers can now bid multiple times in a row but the display is duplicated in the 'View All Bids' display. The duplicate is something like 1.00 if the buyers bid 1000. It is saved correctly in the database, though. So, basically, if the buyer bids 1000, the display in the 'View All Bids' is P1.00 and P1000. This is a bug that needs to be fixed some other time.